backend/api/requirements.txt
Flask==3.1.1
SQLAlchemy==2.0.28
Flask-SQLAlchemy==3.1.1
Flask-Alembic==3.1.1
Flask-Marshmallow>=0.15.0
marshmallow-sqlalchemy>=0.29.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
Flask-JWT-Extended>=4.6.0
